modificacion en perfilCliente

+++ backend/app/Http/Controllers/PerfilClientesController.php
@@ -3,44 +3,56 @@
 namespace App\Http\Controllers;
 
 use App\PerfilCliente;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 
-class PerfilClientesController extends Controller {
-
-    public function store(Request $request) {
+class PerfilClientesController extends Controller
+{
+    public function store(Request $request)
+    {
 
         $this->validate($request, [
-            'nombreComercio'   => 'required',
-            'nombre'           => 'required',
-            'apellido'         => 'required',
-            'documento'        => 'required',
-            'correo'           => 'required',
-            'telefono'         => 'required',
-            'celular'          => 'required',
-            'domicilioEntrega' => 'required',
-            'facturacion'      => 'required',
+            'nombreComercio'     => 'required',
+            'nombre'             => 'required',
+            'apellido'           => 'required',
+            'documento_dni'      => 'required',
+            'documento_otro'     => 'required',
+            'correo'             => 'required',
+            'telefono'           => 'required',
+            'celular'            => 'required',
+            'fk_idPerfilCliente' => 'required',
 
         ], [
-            'nombreComercio.required'   => 'El campo es requerido',
-            'nombre.required'           => 'El campo es requerido',
-            'apellido.required'         => 'El campo es requerido',
-            'documento.required'        => 'El campo es requerido',
-            'correo.required'           => 'El campo es requerido',
-            'telefono.required'         => 'El campo es requerido',
-            'celular.required'          => 'El campo es requerido',
-            'domicilioEntrega.required' => 'El campo es requerido',
-            'facturacion.required'      => 'El campo es requerido',
+            'nombreComercio.required'     => 'El campo es requerido',
+            'nombre.required'             => 'El campo es requerido',
+            'apellido.required'           => 'El campo es requerido',
+            'documento_dni.required'      => 'El campo es requerido',
+            'documento_otro.required'     => 'El campo es requerido',
+            'correo.required'             => 'El campo es requerido',
+            'telefono.required'           => 'El campo es requerido',
+            'celular.required'            => 'El campo es requerido',
+            'fk_idPerfilCliente.required' => 'El campo es requerido',
         ]);
 
         DB::beginTransaction();
+        $user = User::find($request->fk_idPerfilCliente);
+
+        if (is_null($user)) {
+            $response = [
+                'msj' => 'El usuario (fk_idPerfilCliente) no existe',
+            ];
+
+            return response()->json($response, 404);
+        }
 
         try {
 
             $perfilcliente = new PerfilCliente($request->all());
 
             $perfilcliente->save();
+            $perfilcliente->user;
 
             $response = [
                 'msj'  => 'Perfil creado exitosamente',
@@ -60,17 +72,27 @@ class PerfilClientesController extends Controller {
         }
     }
 
-    public function update(Request $request, $idPerfilCliente) {
+    public function update(Request $request, $idPerfilCliente)
+    {
 
         DB::beginTransaction();
 
         try {
-            $pefilCliente = PerfilCliente::findOrFail($idPerfilCliente);
+
+            $pefilCliente = PerfilCliente::find($idPerfilCliente);
+
+            if (is_null($pefilCliente)) {
+                $response = [
+                    'msj' => 'El Perfil del cliente no existe (idPerfilCliente) no existe',
+                ];
+
+                return response()->json($response, 404);
+            }
 
             $pefilCliente->update($request->all());
 
             $response = [
-                'msj'  => 'Info del Perfil actulizada',
+                'msj'    => 'Info del Perfil actulizada',
                 'perfil' => $pefilCliente,
             ];
 
@@ -88,16 +110,26 @@ class PerfilClientesController extends Controller {
         }
     }
 
-    public function destroy($idPerfilCliente) {
+    public function destroy($idPerfilCliente)
+    {
 
         DB::beginTransaction();
 
+        $pefilCliente = PerfilCliente::find($idPerfilCliente);
+
+        if (is_null($pefilCliente)) {
+            $response = [
+                'msj' => 'El Perfil del cliente no existe (idPerfilCliente) no existe',
+            ];
+
+            return response()->json($response, 404);
+        }
         try {
             $pefilCliente = PerfilCliente::findOrFail($idPerfilCliente);
             $pefilCliente->delete();
 
             $response = [
-                'msj'  => 'Perfil eliminado Correctamente',
+                'msj' => 'Perfil eliminado Correctamente',
             ];
 
             DB::commit();
@@ -111,5 +143,17 @@ class PerfilClientesController extends Controller {
                 'message' => 'Ha ocurrido un error al tratar de eliminar los datos.',
             ], 500);
         }
+    }
+
+    public function listar()
+    {
+        $perfil_cliente = PerfilCliente::with('user')->get();
+
+        $response = [
+            'msj'      => 'Lista de perfiles',
+            'perfiles' => $perfil_cliente,
+        ];
+
+        return response()->json($response, 201);
     }
 }
